requirements.txt
PyQt5==5.15.9
Pillow==10.0.0
opencv-python==4.8.1.78
numpy==1.24.4
pytesseract==0.3.10
requests==2.31.0
pywin32==306 